Moved spawning enemy logic out of main
