+++ Main.tscn
@@ -1,14 +1,12 @@
 [gd_scene load_steps=6 format=3 uid="uid://be3yurhibr6t0"]
 
 [ext_resource type="PackedScene" uid="uid://h2iu1fkuw61u" path="res://Scenes/PlayerController/PlayerController.tscn" id="1_bcck4"]
-[ext_resource type="Script" path="res://Main.cs" id="1_lj5n1"]
+[ext_resource type="Script" path="res://TestForSpawningEnemy.cs" id="1_i1bxj"]
 [ext_resource type="PackedScene" uid="uid://oe1ffghf4nah" path="res://Scenes/Enemies/Enemy.tscn" id="2_56tu2"]
 [ext_resource type="PackedScene" uid="uid://ubc5cusrcpq7" path="res://Scenes/Static/Couch.tscn" id="3_fnfmp"]
 [ext_resource type="PackedScene" uid="uid://bihv5yjq7muuh" path="res://Scenes/Static/SicilianHouse.tscn" id="3_ie86o"]
 
 [node name="Main" type="Node"]
-script = ExtResource("1_lj5n1")
-EnemyScene = ExtResource("2_56tu2")
 
 [node name="Camera2D" type="Camera2D" parent="."]
 position = Vector2(288, 164)
@@ -17,6 +15,8 @@ zoom = Vector2(2, 2)
 [node name="GameContainer" type="Node2D" parent="."]
 z_as_relative = false
 y_sort_enabled = true
+script = ExtResource("1_i1bxj")
+EnemyScene = ExtResource("2_56tu2")
 
 [node name="PlayerController" parent="GameContainer" instance=ExtResource("1_bcck4")]
 z_as_relative = false
@@ -29,7 +29,7 @@ position = Vector2(165, 173)
 [node name="Couch" parent="GameContainer" instance=ExtResource("3_fnfmp")]
 position = Vector2(144, 242)
 
-[node name="EnemyTimer" type="Timer" parent="."]
+[node name="EnemyTimer" type="Timer" parent="GameContainer"]
 one_shot = true
 
-[connection signal="timeout" from="EnemyTimer" to="." method="OnEnemyTimerTimeout"]
+[connection signal="timeout" from="GameContainer/EnemyTimer" to="GameContainer" method="OnEnemyTimerTimeout"]
